requirements.txt
fastapi==0.117.1
uvicorn==0.35.0
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-dotenv==1.1.1
langdetect==1.0.9
openai==1.107.2
Pillow==11.3.0
pytesseract==0.3.13
